+++ src/main/java/org/example/repository/BankRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 public interface BankRepository extends JpaRepository<Bank,Long> {
     Optional<Bank> getFirstByName(String name);
 
-    Optional<Bank> getFirstByBikCode(Long bikCode);
+    Optional<Bank> getFirstByBikCode(Integer bikCode);
 }
